clientconfig: Deprioritize Environment Variables
This commit will make clientconfig only use an environment variable if
an authentication setting was not previously set. This is to fix issues
where an application consuming clientconfig has already determined that
environment variables are not required.

This behavior matches Python-based applications since specifying an
authentication setting on the command-line will take priority of an
environment variable.