Don't fail if USER is not defined
There's a way to workaround this issue by setting OS_CLIENT_CONFIG_FILE
to the right path, it's still not ok to fail when a USER is not defined
in the system. This can happen in various scenarios, including running
within a container.

OS_CLIENT_CONFIG_FILE should be a feature and not a workaround for this
problem. If USER is not defined or the `~/.config/openstack` dir
doesn't exist, we should just move on.