Fixed gnocchi metrics and resources list pager
In the current implementation, pager type of gnocchi metrics and resources list
is SinglePageBase, which is wrong. In Gnocchi, query results in one page
are limited to max_limit value set in the configuration file (default 1000).
With the current implementation, AllPages() method can only show the first
page of query result (the first 1000 metrics, for example).

The pager type should be MarkerPageBase and LastMarker() method needs to
be implemented.