Consider TLS options when we create a service Client
Now, when we create a service client with gophercloud/utils, TLS options[1]
are ignored, which leads to the fact that we can't work with
self-signed certificates out of the box. Also it creates inconsistencies
with python-openstackclient, where these options matter.

To remove the discrepancy this commit implements full TLS support.
The work is based on the Terraform client[2], and now that code is
shared between the clients.

[1] https://github.com/gophercloud/utils/blob/88c6e9e533261350ec733e2e2147a1d3f8eb62de/openstack/clientconfig/results.go#L39-L49
[2] https://github.com/gophercloud/utils/blob/88c6e9e533261350ec733e2e2147a1d3f8eb62de/terraform/auth/config.go#L144-L179