openstack/clientconfig/requests.go: make sure tls config is generated before creating authenticated client (#140)
Previously the function would call AuthenticatedClient function which would try authenitcating without the appropiate
tls configuration. Since the function is public, changing it's behavior would be not ideal, so I have moved the 2 steps
of creating provider client and authenticating it into this functions itself. So now the function creates a provider client,
updates the HTTPClient to correct tls configuration and uses that to authenticate.