Terraform: use reference to MutexKV in Config
We don't want to have a copy of the `sync.Locker` in the
`Config` since lock value shold not be copied.

Usually `Config` used as a singleton in different Terraform provider
resources and all of them should use a single `sync.Locker` instead of
different copies.

This commit changes `MutexKV` field of the `Config` to be a reference
instead of a copy and updates `MutexKV` tests to check that `MutexKV` is
used in a proper way.