clientconfig: Ignore OS_CLOUD if explicit cloud config provided
The current behavior of the 'GetCloudFromYAML' and 'AuthOptions'
utilities is confusing and the source of much pulled hair. As discussed
in the linked bug, the equivalent Python libraries, openstacksdk and
(legacy) os_client_config, both default to reading 'OS_CLOUD' only if
explicit cloud config has not been provided. However, we're actually
overriding what the user provided if 'OS_CLOUD' is set. This is clearly
incorrect. Correct things to prefer user-provide cloud configuration,
falling back to the environment variable only if this is not provided.

While we're here, correctly handle a nil 'opts' arguments in the
'GetCloudFromYAML' function. Previously this was done ad-hoc and was
incomplete.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Closes: #164