Replace go-homedir with stdlib calls
The purpose of github.com/mitchellh/go-homedir was to expand the user
home directory without CGO. Recent versions of `os/user` can retrieve
the user home directory without CGO and thus enable cross-compilation.

This change does not add the missing functionality of expanding
non-current-user home directories in a path (e.g. `~user2/dir/subdir`).
The reason for this is that the input of `pathOrContents` could be a
blob and parsing every byte slice matching `^~.*\/.*` as a path
considerably increases the chances that a binary blob is mistakenly
parsed as a path.